fix some mock
