+++ dm/worker/worker_test.go
@@ -15,11 +15,13 @@ package worker
 
 import (
 	"context"
+	"database/sql"
 	"fmt"
 	"sync"
 	"sync/atomic"
 	"time"
 
+	"github.com/DATA-DOG/go-sqlmock"
 	. "github.com/pingcap/check"
 	"github.com/pingcap/failpoint"
 	"github.com/tikv/pd/pkg/tempurl"
@@ -28,12 +30,36 @@ import (
 	"github.com/pingcap/dm/dm/config"
 	"github.com/pingcap/dm/dm/pb"
 	"github.com/pingcap/dm/dm/unit"
+	"github.com/pingcap/dm/pkg/conn"
 	"github.com/pingcap/dm/pkg/ha"
 	"github.com/pingcap/dm/pkg/log"
 	"github.com/pingcap/dm/pkg/utils"
 )
 
 var emptyWorkerStatusInfoJSONLength = 25
+
+type mockDBProvider struct {
+	db *sql.DB
+}
+
+// Apply will build BaseDB with DBConfig.
+func (d *mockDBProvider) Apply(config config.DBConfig) (*conn.BaseDB, error) {
+	return conn.NewBaseDB(d.db, func() {}), nil
+}
+
+func initMockDB(c *C) sqlmock.Sqlmock {
+	db, mock, err := sqlmock.New()
+	c.Assert(err, IsNil)
+	conn.DefaultDBProvider = &mockDBProvider{db: db}
+	return mock
+}
+
+func mockShowMasterStatus(mockDB sqlmock.Sqlmock) {
+	rows := mockDB.NewRows([]string{"File", "Position", "Binlog_Do_DB", "Binlog_Ignore_DB", "Executed_Gtid_Set"}).AddRow(
+		"mysql-bin.000009", 11232, nil, nil, "074be7f4-f0f1-11ea-95bd-0242ac120002:1-699",
+	)
+	mockDB.ExpectQuery(`SHOW MASTER STATUS`).WillReturnRows(rows)
+}
 
 func (t *testServer) testWorker(c *C) {
 	cfg := loadSourceConfigWithoutPassword(c)
@@ -164,7 +190,6 @@ func (t *testServer2) TestTaskAutoResume(c *C) {
 	defer failpoint.Disable("github.com/pingcap/dm/dumpling/dumpUnitProcessWithError")
 
 	s := NewServer(cfg)
-
 	defer s.Close()
 	go func() {
 		c.Assert(s.Start(), IsNil)
@@ -184,6 +209,7 @@ func (t *testServer2) TestTaskAutoResume(c *C) {
 	var subtaskCfg config.SubTaskConfig
 	c.Assert(subtaskCfg.DecodeFile("./subtask.toml", true), IsNil)
 	c.Assert(err, IsNil)
+	subtaskCfg.Mode = "full"
 	c.Assert(s.getWorker(true).StartSubTask(&subtaskCfg, pb.Stage_Running, true), IsNil)
 
 	// check task in paused state
@@ -206,9 +232,6 @@ func (t *testServer2) TestTaskAutoResume(c *C) {
 		rtsc.Close()
 		rtsc.Close()
 	}()
-
-	mockDB := initMockDB(c)
-	mockShowMasterStatus(mockDB)
 
 	// check task will be auto resumed
 	c.Assert(utils.WaitSomething(10, 100*time.Millisecond, func() bool {
